# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API",
    "version": "1.0.0",
    "description": "API for managing employee tasks and work plans"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://workplan-api.onrender.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "tools": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "time": {
            "type": "string",
            "format": "time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "constraints": {
            "type": "string"
          }
        }
      },
      "Plan": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ]
          },
          "task": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/ps/dashboard/departments/unit": {
      "parameters": [],
      "post": {
        "summary": "Get Staff in a Unit",
        "parameters": [],
        "responses": {},
        "tags": [
          "PS"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "department_id": {
                    "type": "string",
                    "format": "utc-millisec"
                  },
                  "unit_id": {
                    "type": "string",
                    "format": "utc-millisec"
                  },
                  "week": {
                    "type": "string"
                  },
                  "month": {
                    "type": "string"
                  },
                  "role_filter": {
                    "type": "string"
                  },
                  "status_filter": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "department_id": "3",
                "unit_id": "2",
                "week": "WEEK_1",
                "month": "AUGUST",
                "role_filter": "",
                "status_filter": ""
              }
            }
          }
        }
      }
    },
...
    "/api/hou/plan/create": {
      "post": {
        "tags": [
          "HOU Actions"
        ],
        "summary": "Create plan",
        "description": "Create a new work plan with tasks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "month": {
                    "type": "string",
                    "enum": [
                      "JANUARY",
                      "FEBRUARY",
                      "MARCH",
                      "APRIL",
                      "MAY",
                      "JUNE",
                      "JULY",
                      "AUGUST",
                      "SEPTEMBER",
                      "OCTOBER",
                      "NOVEMBER",
                      "DECEMBER"
                    ]
                  },
                  "week": {
                    "type": "string",
                    "enum": [
                      "WEEK_1",
                      "WEEK_2",
                      "WEEK_3",
                      "WEEK_4"
                    ]
                  },
                  "plans": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Plan"
                    }
                  }
                },
                "required": [
                  "month",
                  "week",
                  "plans"
                ]
              },
              "example": {
                "month": "AUGUST",
                "week": "WEEK_1",
                "plans": [
                  {
                    "day": "MONDAY",
                    "task": [
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "04/07/2025"
                      },
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "04/07/2025"
                      }
                    ]
                  },
                  {
                    "day": "TUESDAY",
                    "task": [
                      {
                        "title": "Team meeting",
                        "notes": "Discuss project roadmap and blockers",
                        "tools": "Zoom",
                        "priority": "MEDIUM",
                        "time": "11:00",
                        "status": "IN_PROGRESS",
                        "constraints": "All team members must be available, requires stable internet connection",
                        "date": "05/07/2025"
                      },
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "05/07/2025"
                      }
                    ]
                  },
                  {
                    "day": "WEDNESDAY",
                    "task": [
                      {
                        "title": "Code review",
                        "notes": "Review PR #42 from frontend team",
                        "tools": "GitHub",
                        "priority": "LOW",
                        "time": "14:00",
                        "status": "PENDING",
                        "constraints": "PR must be submitted and ready for review, requires development environment access",
                        "date": "05/07/2025"
                      },
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "05/07/2025"
                      }
                    ]
                  },
                  {
                    "day": "THURSDAY",
                    "task": [
                      {
                        "title": "Write documentation",
                        "notes": "Update API endpoints in Notion",
                        "tools": "Notion",
                        "priority": "MEDIUM",
                        "time": "10:00",
                        "status": "PENDING",
                        "constraints": "API changes must be finalized, requires Notion workspace access",
                        "date": "05/07/2025"
                      },
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "05/07/2025"
                      }
                    ]
                  },
                  {
                    "day": "FRIDAY",
                    "task": [
                      {
                        "title": "Deploy new version",
                        "notes": "Push v1.2 to production servers",
                        "tools": "AWS CLI",
                        "priority": "HIGH",
                        "time": "16:00",
                        "status": "PENDING",
                        "constraints": "Code must pass all tests, requires production deployment permissions, maintenance window needed",
                        "date": "05/07/2025"
                      },
                      {
                        "title": "Finish report",
                        "notes": "Include Q3 data and finalize charts",
                        "priority": "HIGH",
                        "time": "09:00",
                        "status": "PENDING",
                        "date": "05/07/2025"
                      }
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hou/plan/report": {
      "post": {
        "tags": [
          "HOU Actions"
        ],
        "summary": "Get Report",
        "description": "Generate report with filters (Note: GET with requestBody is not standard, consider using POST)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "week": {
                    "type": "string",
                    "enum": [
                      "WEEK_1",
                      "WEEK_2",
                      "WEEK_3",
                      "WEEK_4"
                    ]
                  },
                  "month": {
                    "type": "string",
                    "enum": [
                      "JANUARY",
                      "FEBRUARY",
                      "MARCH",
                      "APRIL",
                      "MAY",
                      "JUNE",
                      "JULY",
                      "AUGUST",
                      "SEPTEMBER",
                      "OCTOBER",
                      "NOVEMBER",
                      "DECEMBER"
                    ]
                  }
                }
              },
              "example": {
                "week": "WEEK_2",
                "month": "AUGUST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": {}
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}